--- packages/commonwealth/client/scripts/state/api/token/index.ts
@@ -1,4 +1,0 @@
-import useCreateTokenMutation from './createToken';
-import useFetchTokensQuery from './fetchTokens';
-
-export { useCreateTokenMutation, useFetchTokensQuery };
