+++ packages/commonwealth/client/scripts/state/api/tokens/index.ts
@@ -1,3 +1,11 @@
+import useCreateTokenMutation from './createToken';
+import useFetchTokensQuery from './fetchTokens';
+import useTokenBalanceQuery from './getTokenBalance';
 import useTokenMetadataQuery from './getTokenMetadata';
 
-export { useTokenMetadataQuery };
+export {
+  useCreateTokenMutation,
+  useFetchTokensQuery,
+  useTokenBalanceQuery,
+  useTokenMetadataQuery,
+};
